+++ src/Entities/Player.h
@@ -28,7 +28,7 @@ class Player : public Entity
          if (index == ID("rndr"))
             return 40;
          else
-            return (int)this;
+            return (uintptr_t)this;
       }
 
       const char*
